Store the command-line options passed by the user in a separate `@options` instance variable
This commit is inspired by issue #79 and the solution suggested by @jbraeuer
(https://github.com/davetron5000/gli/issues/79#issuecomment-9913521).
The issue addresses the problem of multiple config files, which is a
feature that GLI does not have yet. The suggested solution works (with a
little bit of tweaking) but the options passed by the user on the
command-line are always overwritten by the options in the config file
that was parsed last. By storing the command-line options, one can avoid
this problem.